Move createTestDictionaryArchive, remove yomichan-test.js
