+++ test/test-database.js
@@ -19,8 +19,7 @@ const fs = require('fs');
 const url = require('url');
 const path = require('path');
 const assert = require('assert');
-const {JSZip} = require('../dev/yomichan-util');
-const {createTestDictionaryArchive} = require('./yomichan-test');
+const {JSZip, createDictionaryArchive} = require('../dev/yomichan-util');
 const {VM} = require('./yomichan-vm');
 require('fake-indexeddb/auto');
 
@@ -123,6 +122,12 @@ vm.execute([
 ]);
 const DictionaryImporter = vm.get('DictionaryImporter');
 const DictionaryDatabase = vm.get('DictionaryDatabase');
+
+
+function createTestDictionaryArchive(dictionary, dictionaryName) {
+    const dictionaryDirectory = path.join(__dirname, 'data', 'dictionaries', dictionary);
+    return createDictionaryArchive(dictionaryDirectory, dictionaryName);
+}
 
 
 function countTermsWithExpression(terms, expression) {
